+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 ## Unreleased
 - [**BC**] Require `CurrentEnvironment` in all kafka applications
 - Add `ApplicationStatus`
+- Update dependencies
+- [**BC**] Use `Giraffe` and `Saturn` instead of `Suave` and use internal `WebServer` instead of Metrics webserver
+- [**BC**] Change `showMetricsOn` to `showMetrics` use default `/metrics` route statically
+- [**BC**] Change `addRoute` to `addHttpHandler`
+- [**BC**] Change `useGitCommit` to `useGit`
+- [**BC**] Remove `gitCommit` from `environment` computation expression
 
 ## 16.0.0 - 2021-12-17
 - [**BC**] Require `Instance` in common `LoggerFactory`